Make the analyzer useful
The analyzer now only needs to do its work once, and macros are fully
expanded as part of the process. Some syntax errors and attempts to
access unbound variables are also identified during analysis, but no
optimization is performed.

To make analysis easier, I've removed the macro special form and added a
few kludges:

- defmacro is now a special form
- def and defmacro can only be used as top-level forms
- set! can only be used on locals

The Primitive and Macro types are gone, expand-all and exec are in.